i/requirements.txt
asgiref==3.4.1
astroid==2.8.0
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
crypto==1.4.1
Django==3.2.7
django-auth-ldap==4.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
lazy-object-proxy==1.7.1
multi-key-dict==2.0.3
mysqlclient==2.1.0
Naked==0.1.31
pbr==5.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-jenkins==1.7.0
python-ldap==3.4.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.3
typing_extensions==4.0.1
urllib3==1.26.8
wrapt==1.12.1
shortuuid==1.0.8
grpcio==1.41.0
grpcio-tools==1.41.0
kafka-python==2.0.2
func-timeout==4.3.5
